Fixed getdailyRedemption point crash, and added some test cases for withdraw more than the current balance

+++ frontend/AndroidCasinoUser/app/src/main/java/com/davidzhang/androidcasinouser/UserProfile.java
@@ -119,7 +119,7 @@ public class UserProfile extends ThemedActivity {
             public void call(Object... args) {
                 Log.d(TAG, "received new balance details");
                 if (args[0] != null) {
-                    double newbalance = (double) args[0];
+                    Integer newbalance = (Integer) args[0];
 
                     runOnUiThread(new Runnable() {
                         @Override
